HACK fractional scaling

+++ terminal.h
@@ -445,7 +445,7 @@ struct terminal {
         int fd;
     } blink;
 
-    int scale;
+    double scale;
     int width;  /* pixels */
     int height; /* pixels */
     int stashed_width;
